better to upgrade like this
otherwise you'll get errors on older UNRAID installations not being able to upgrade to 2.8.0 (and not having that file in the first place), also you'll always trigger the symlink check because /etc/nut/ will never exist at that stage (with the package not being installed at the time of checking). also better for diagnostics to know which package exactly the user has installed by not masquerading with a common filename.